Update bill of materials to final design.
There are two differences to the real prototype:

* The metal posts (zinc coated) are not made of 2 angle beams,
because it turned out difficult to bend an entire 90 degree angle
beam to 120 degree (due to 1.25m length and lack of proper machines).

Therefore instead a piece of a bit less than 3cm length has been cut
and used instead for every bolting required to interconnect the side
stabilizers (60degree angle to main/long siding).

* Instead of M10 (approximate 10mm outer thread diameter) bolts, M8
(8mm) have been used. It's discouraged to use the 10cm bolts because of
less material remaining at the beam steel due to the close placement to
the edge (often 1cm or 1.4cm, see .blend 3D model file).